backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
aiosqlite==0.20.0
httpx==0.27.0
anthropic>=0.30.0
playwright==1.44.0
python-whois==0.9.4
googlemaps==4.10.0
apscheduler==3.10.4
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.2.2
fake-useragent==1.5.1
pydantic==2.7.1
pydantic-settings==2.3.0
python-multipart==0.0.9
websockets==12.0
openpyxl==3.1.4
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

